--- scheduler-service/src/main/java/com/codeaddi/scheduler_service/model/repository/PastSessionsRepository.java
@@ -1,17 +1,0 @@
-package com.codeaddi.scheduler_service.model.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface PastSessionsRepository extends JpaRepository<PastSession, Long> {
-
-
-
-}
-
