+++ scheduler-service/src/main/java/com/codeaddi/scheduler_service/model/repository/sessions/SessionRepository.java
@@ -1,4 +1,4 @@
-package com.codeaddi.scheduler_service.model.repository;
+package com.codeaddi.scheduler_service.model.repository.sessions;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
